+++ xonsh/built_ins.py
@@ -13,6 +13,7 @@ import atexit
 import signal
 import inspect
 import builtins
+import tempfile
 from glob import glob, iglob
 from subprocess import Popen, PIPE, STDOUT, CalledProcessError
 from contextlib import contextmanager
@@ -20,14 +21,15 @@ from collections import Sequence, MutableMapping, Iterable
 
 from xonsh.tools import (
     suggest_commands, XonshError, ON_POSIX, ON_WINDOWS, string_types,
-    expandvars,
+    expandvars, CommandsCache
 )
 from xonsh.inspectors import Inspector
 from xonsh.environ import Env, default_env, locate_binary
 from xonsh.aliases import DEFAULT_ALIASES
 from xonsh.jobs import add_job, wait_for_active_job
-from xonsh.proc import (ProcProxy, SimpleProcProxy, ForegroundProcProxy, 
-    SimpleForegroundProcProxy,TeePTYProc)
+from xonsh.proc import (ProcProxy, SimpleProcProxy, ForegroundProcProxy,
+                        SimpleForegroundProcProxy, TeePTYProc,
+                        CompletedCommand, HiddenCompletedCommand)
 from xonsh.history import History
 from xonsh.foreign_shells import load_foreign_aliases
 
@@ -45,6 +47,7 @@ def resetting_signal_handle(sig, f):
     once the new handle is finished.
     """
     oldh = signal.getsignal(sig)
+
     def newh(s=None, frame=None):
         f(s, frame)
         signal.signal(sig, oldh)
@@ -329,10 +332,10 @@ def get_script_subproc_command(fname, args):
         raise PermissionError
 
     if ON_POSIX and not os.access(fname, os.R_OK):
-        # on some systems, some importnat programs (e.g. sudo) will have execute
-        # permissions but not read/write permisions. This enables things with the SUID
-        # set to be run. Needs to come before _is_binary() is called, because that
-        # function tries to read the file.
+        # on some systems, some importnat programs (e.g. sudo) will have
+        # execute permissions but not read/write permisions. This enables
+        # things with the SUID set to be run. Needs to come before _is_binary()
+        # is called, because that function tries to read the file.
         return [fname] + args
     elif _is_binary(fname):
         # if the file is a binary, we should call it directly
@@ -406,7 +409,7 @@ def _redirect_io(streams, r, loc=None):
     if r.replace('&', '') in _E2O_MAP:
         if 'stderr' in streams:
             raise XonshError('Multiple redirects for stderr')
-        streams['stderr'] = STDOUT
+        streams['stderr'] = ('<stdout>', 'a', STDOUT)
         return
 
     orig, mode, dest = _REDIR_REGEX.match(r).groups()
@@ -433,7 +436,7 @@ def _redirect_io(streams, r, loc=None):
         elif 'stdin' in streams:
             raise XonshError('Multiple inputs for stdin')
         else:
-            streams['stdin'] = _open(loc, mode)
+            streams['stdin'] = (loc, 'r', _open(loc, mode))
     elif mode in _WRITE_MODES:
         if orig in _REDIR_ALL:
             if 'stderr' in streams:
@@ -463,13 +466,13 @@ def _redirect_io(streams, r, loc=None):
 
         f = _open(loc, mode)
         for t in targets:
-            streams[t] = f
+            streams[t] = (loc, mode, f)
 
     else:
         raise XonshError('Unrecognized redirection command: {}'.format(r))
 
 
-def run_subproc(cmds, captured=True):
+def run_subproc(cmds, captured=False):
     """Runs a subprocess, in its many forms. This takes a list of 'commands,'
     which may be a list of command line arguments or a string, representing
     a special connecting character.  For example::
@@ -484,6 +487,7 @@ def run_subproc(cmds, captured=True):
     """
     global ENV
     background = False
+    procinfo = {}
     if cmds[-1] == '&':
         background = True
         cmds = cmds[:-1]
@@ -491,7 +495,9 @@ def run_subproc(cmds, captured=True):
     last_cmd = len(cmds) - 1
     procs = []
     prev_proc = None
+    _capture_streams = captured in {'stdout', 'object'}
     for ix, cmd in enumerate(cmds):
+        procinfo['args'] = list(cmd)
         stdin = None
         stderr = None
         if isinstance(cmd, string_types):
@@ -513,15 +519,17 @@ def run_subproc(cmds, captured=True):
         if 'stdin' in streams:
             if prev_proc is not None:
                 raise XonshError('Multiple inputs for stdin')
-            stdin = streams['stdin']
+            stdin = streams['stdin'][-1]
+            procinfo['stdin_redirect'] = streams['stdin'][:-1]
         elif prev_proc is not None:
             stdin = prev_proc.stdout
         # set standard output
         if 'stdout' in streams:
             if ix != last_cmd:
                 raise XonshError('Multiple redirects for stdout')
-            stdout = streams['stdout']
-        elif captured or ix != last_cmd:
+            stdout = streams['stdout'][-1]
+            procinfo['stdout_redirect'] = streams['stdout'][:-1]
+        elif _capture_streams or ix != last_cmd:
             stdout = PIPE
         elif builtins.__xonsh_stdout_uncaptured__ is not None:
             stdout = builtins.__xonsh_stdout_uncaptured__
@@ -529,16 +537,20 @@ def run_subproc(cmds, captured=True):
             stdout = None
         # set standard error
         if 'stderr' in streams:
-            stderr = streams['stderr']
+            stderr = streams['stderr'][-1]
+            procinfo['stderr_redirect'] = streams['stderr'][:-1]
+        elif captured == 'object':
+            stderr = PIPE
         elif builtins.__xonsh_stderr_uncaptured__ is not None:
             stderr = builtins.__xonsh_stderr_uncaptured__
-        uninew = (ix == last_cmd) and (not captured)
+        uninew = (ix == last_cmd) and (not _capture_streams)
         alias = builtins.aliases.get(cmd[0], None)
-        if (alias is None
-            and builtins.__xonsh_env__.get('AUTO_CD')
-            and len(cmd) == 1
-            and os.path.isdir(cmd[0])
-            and locate_binary(cmd[0]) is None):
+        procinfo['alias'] = alias
+        if (alias is None and
+                builtins.__xonsh_env__.get('AUTO_CD') and
+                len(cmd) == 1 and
+                os.path.isdir(cmd[0]) and
+                locate_binary(cmd[0]) is None):
             cmd.insert(0, 'cd')
             alias = builtins.aliases.get('cd', None)
 
@@ -556,6 +568,20 @@ def run_subproc(cmds, captured=True):
                 except PermissionError:
                     e = 'xonsh: subprocess mode: permission denied: {0}'
                     raise XonshError(e.format(cmd[0]))
+        _stdin_file = None
+        if (captured == 'object' and
+                stdin is not None and
+                __xonsh_env__['XONSH_STORE_STDIN'] and
+                'cat' in __xonsh_commands_cache__ and
+                'tee' in __xonsh_commands_cache__):
+            _stdin_file = tempfile.NamedTemporaryFile()
+            cproc = Popen(['cat'],
+                          stdin=stdin,
+                          stdout=PIPE)
+            tproc = Popen(['tee', _stdin_file.name],
+                          stdin=cproc.stdout,
+                          stdout=PIPE)
+            stdin = tproc.stdout
         if callable(aliased_cmd):
             prev_is_proxy = True
             bgable = getattr(aliased_cmd, '__xonsh_backgroundable__', True)
@@ -572,8 +598,9 @@ def run_subproc(cmds, captured=True):
                        universal_newlines=uninew)
         else:
             prev_is_proxy = False
-            usetee = (stdout is None) and (not background) and \
-                     ENV.get('XONSH_STORE_STDOUT', False)
+            usetee = ((stdout is None) and
+                      (not background) and
+                      ENV.get('XONSH_STORE_STDOUT', False))
             cls = TeePTYProc if usetee else Popen
             subproc_kwargs = {}
             if ON_POSIX and cls is Popen:
@@ -610,7 +637,9 @@ def run_subproc(cmds, captured=True):
             'obj': prev_proc,
             'bg': background
         })
-    if ENV.get('XONSH_INTERACTIVE') and not ENV.get('XONSH_STORE_STDOUT'):
+    if (ENV.get('XONSH_INTERACTIVE') and
+            not ENV.get('XONSH_STORE_STDOUT') and
+            not _capture_streams):
         # set title here to get current command running
         try:
             builtins.__xonsh_shell__.settitle()
@@ -628,7 +657,7 @@ def run_subproc(cmds, captured=True):
     if write_target is None:
         if prev_proc.stdout not in (None, sys.stdout):
             output = prev_proc.stdout.read()
-        if captured:
+        if _capture_streams:
             # to get proper encoding from Popen, we have to
             # use a byte stream and then implement universal_newlines here
             output = output.decode(encoding=ENV.get('XONSH_ENCODING'),
@@ -636,17 +665,55 @@ def run_subproc(cmds, captured=True):
             output = output.replace('\r\n', '\n')
         else:
             hist.last_cmd_out = output
-    if not prev_is_proxy and hist.last_cmd_rtn > 0 and ENV.get('RAISE_SUBPROC_ERROR'):
+        if (captured == 'object' and
+                prev_proc.stderr not in {None, sys.stderr}):
+            errout = prev_proc.stderr.read()
+            errout = errout.decode(encoding=ENV.get('XONSH_ENCODING'),
+                                   errors=ENV.get('XONSH_ENCODING_ERRORS'))
+            errout = errout.replace('\r\n', '\n')
+            procinfo['stderr'] = errout
+
+    if (not prev_is_proxy and
+            hist.last_cmd_rtn > 0 and
+            ENV.get('RAISE_SUBPROC_ERROR')):
         raise CalledProcessError(hist.last_cmd_rtn, aliased_cmd, output=output)
-    if captured:
+    if captured == 'stdout':
         return output
+    elif captured is not False:
+        procinfo['pid'] = prev_proc.pid
+        procinfo['returncode'] = prev_proc.returncode
+        if captured == 'object':
+            procinfo['stdout'] = output
+            if _stdin_file is not None:
+                _stdin_file.seek(0)
+                procinfo['stdin'] = _stdin_file.read().decode()
+                _stdin_file.close()
+            return CompletedCommand(**procinfo)
+        else:
+            return HiddenCompletedCommand(**procinfo)
 
 
-def subproc_captured(*cmds):
+def subproc_captured_stdout(*cmds):
     """Runs a subprocess, capturing the output. Returns the stdout
     that was produced as a str.
     """
-    return run_subproc(cmds, captured=True)
+    return run_subproc(cmds, captured='stdout')
+
+
+def subproc_captured_object(*cmds):
+    """
+    Runs a subprocess, capturing the output. Returns an instance of
+    ``CompletedCommand`` representing the completed command.
+    """
+    return run_subproc(cmds, captured='object')
+
+
+def subproc_captured_hiddenobject(*cmds):
+    """
+    Runs a subprocess, capturing the output. Returns an instance of
+    ``HiddenCompletedCommand`` representing the completed command.
+    """
+    return run_subproc(cmds, captured='hiddenobject')
 
 
 def subproc_uncaptured(*cmds):
@@ -689,9 +756,12 @@ def load_builtins(execer=None, config=None):
     if hasattr(builtins, 'quit'):
         builtins.__xonsh_pyquit__ = builtins.quit
         del builtins.quit
-    builtins.__xonsh_subproc_captured__ = subproc_captured
+    builtins.__xonsh_subproc_captured_stdout__ = subproc_captured_stdout
+    builtins.__xonsh_subproc_captured_object__ = subproc_captured_object
+    builtins.__xonsh_subproc_captured_hiddenobject__ = subproc_captured_hiddenobject
     builtins.__xonsh_subproc_uncaptured__ = subproc_uncaptured
     builtins.__xonsh_execer__ = execer
+    builtins.__xonsh_commands_cache__ = CommandsCache()
     builtins.__xonsh_all_jobs__ = {}
     builtins.__xonsh_active_job__ = None
     builtins.__xonsh_ensure_list_of_strs__ = ensure_list_of_strs
@@ -703,13 +773,17 @@ def load_builtins(execer=None, config=None):
     builtins.aliases.update(load_foreign_aliases(issue_warning=False))
     # history needs to be started after env and aliases
     # would be nice to actually include non-detyped versions.
-    builtins.__xonsh_history__ = History(env=ENV.detype(), #aliases=builtins.aliases,
+    builtins.__xonsh_history__ = History(env=ENV.detype(),
                                          ts=[time.time(), None], locked=True)
-    lastflush = lambda s=None, f=None: builtins.__xonsh_history__.flush(at_exit=True)
+    lastflush = _lastflush
     atexit.register(lastflush)
     for sig in AT_EXIT_SIGNALS:
         resetting_signal_handle(sig, lastflush)
     BUILTINS_LOADED = True
+
+
+def _lastflush(s=None, f=None):
+    builtins.__xonsh_history__.flush(at_exit=True)
 
 
 def unload_builtins():
@@ -738,9 +812,12 @@ def unload_builtins():
              '__xonsh_stderr_uncaptured__',
              '__xonsh_pyexit__',
              '__xonsh_pyquit__',
-             '__xonsh_subproc_captured__',
+             '__xonsh_subproc_captured_stdout__',
+             '__xonsh_subproc_captured_object__',
+             '__xonsh_subproc_captured_hiddenobject__',
              '__xonsh_subproc_uncaptured__',
              '__xonsh_execer__',
+             '__xonsh_commands_cache__',
              'evalx',
              'execx',
              'compilex',