+++ xonsh/lexer.py
@@ -218,6 +218,26 @@ def handle_question(state, token, stream):
             yield from handle_token(state, n, stream)
 
 
+def handle_bang(state, token, stream):
+    """
+    Function for generating PLY tokens starting with !
+    """
+    n = next(stream, None)
+    if (n is not None and n.type == tokenize.OP
+            and n.string == '(' and n.start == token.end):
+        state['pymode'].append((False, '!(', ')', token.start))
+        state['last'] = n
+        yield _new_token('BANG_LPAREN', '!(', token.start)
+    elif (n is not None and n.type == tokenize.OP
+            and n.string == '[' and n.start == token.end):
+        state['pymode'].append((False, '![', ']', token.start))
+        state['last'] = n
+        yield _new_token('BANG_LBRACKET', '![', token.start)
+    else:
+        e = 'unexpected single token: !'
+        yield _new_token("ERRORTOKEN", e, token.start)
+
+
 def handle_backtick(state, token, stream):
     """
     Function for generating PLY tokens representing regex globs.
@@ -369,6 +389,7 @@ special_handlers = {
     (tokenize.ERRORTOKEN, '$'): handle_dollar,
     (tokenize.ERRORTOKEN, '`'): handle_backtick,
     (tokenize.ERRORTOKEN, '?'): handle_question,
+    (tokenize.ERRORTOKEN, '!'): handle_bang,
     (tokenize.ERRORTOKEN, ' '): handle_error_space,
 }
 """
@@ -517,6 +538,8 @@ class Lexer(object):
         'DOUBLE_QUESTION',       # ??
         'AT_LPAREN',             # @(
         'DOLLAR_NAME',           # $NAME
+        'BANG_LPAREN',           # !(
+        'BANG_LBRACKET',         # ![
         'DOLLAR_LPAREN',         # $(
         'DOLLAR_LBRACE',         # ${
         'DOLLAR_LBRACKET',       # $[