+++ CHANGELOG.rst
@@ -25,6 +25,9 @@ Current Developments
 * Python mode output is now syntax highlighted if pygments is available.
 * New ``$RIGHT_PROMPT`` environment variable for displaying right-aligned
   text in prompt-toolkit shell.
+* Added ``!(...)`` operator, which returns an object representing the result
+  of running a command.  This object is truthy if the command had a return
+  code of zero, and it is equal (via ``==``) to its return code value.
 * Optional dependency on the win_unicode_console package to enable unicode
   support in cmd.exe on Windows. This can be disabled/enabled with the
   ``$WIN_UNICODE_CONSOLE`` environment variable.
@@ -56,6 +59,8 @@ Current Developments
 **Fixed:**
 
 * Some minor zsh fixes for more platforms and setups.
+* The ``BaseShell.settitle`` method no longer has its commands captured by
+  ``$(...)``
 
 **Security:** None